+++ ui/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - NativeSQL translator: Added support for steps concatenate and fromdate
+- NativeSQL translator: Added support for the percentage step
+- NativeSQL translator: Added support for the concatenate step
 
 ## [0.106.0] - 2023-07-31